Updated code for Gyro and NXT color
Updated the code for Gyro sensor reading, it now accurately returns turning angle and angle change rate (in degrees per second). Also added a function to read from the NXT color sensor, since the EV3 color sensor continues to misbehave, power down, and reset. This is a problem with the sensor itself as it happens with independent implementations of the sensor reading code, and both in C and Python as well as in their own block-based programming software. NXT colour sensor should be preferred as it reliably returns RGB+ambient, no issues with power downs or resets or mode changes.